+++ validation/src/main/java/org/ehrbase/openehr/sdk/validation/webtemplate/DvCountValidator.java
@@ -18,7 +18,6 @@
 package org.ehrbase.openehr.sdk.validation.webtemplate;
 
 import com.nedap.archie.rm.datavalues.quantity.DvCount;
-import java.util.Collections;
 import java.util.List;
 import org.ehrbase.openehr.sdk.validation.ConstraintViolation;
 import org.ehrbase.openehr.sdk.webtemplate.model.WebTemplateNode;
@@ -31,8 +30,6 @@ import org.ehrbase.openehr.sdk.webtemplate.model.WebTemplateNode;
  */
 @SuppressWarnings("unused")
 public class DvCountValidator implements ConstraintValidator<DvCount> {
-
-    private final PrimitiveConstraintValidator validator = new PrimitiveConstraintValidator();
 
     /**
      * {@inheritDoc}
@@ -48,10 +45,10 @@ public class DvCountValidator implements ConstraintValidator<DvCount> {
     @Override
     public List<ConstraintViolation> validate(DvCount dvCount, WebTemplateNode node) {
         if (!WebTemplateValidationUtils.hasInputs(node)) {
-            return Collections.emptyList();
+            return List.of();
         }
 
         var input = WebTemplateValidationUtils.getInputWithType(node, "INTEGER");
-        return validator.validate(node.getAqlPath(), dvCount.getMagnitude(), input);
+        return PrimitiveConstraintValidator.validate(node.getAqlPath(), dvCount.getMagnitude(), input);
     }
 }